+++ packages/agent-templates/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@lobechat/agent-templates",
+  "version": "1.0.0",
+  "private": true,
+  "main": "./src/index.ts"
+}
