+++ packages/agent-templates/src/templates/claw/index.ts
@@ -1,12 +1,6 @@
-/**
- * Claw Policy
- *
- * Sharp, evolving agent with retractable claws that grip onto identity and purpose.
- * Similar to OpenClaw but with structured document loading.
- */
-
 import type { DocumentTemplateSet } from '../index';
 import { AGENT_DOCUMENT } from './agent';
+import { BOOTSTRAP_DOCUMENT } from './bootstrap';
 import { IDENTITY_DOCUMENT } from './identity';
 import { SOUL_DOCUMENT } from './soul';
 
@@ -18,8 +12,8 @@ export const CLAW_POLICY: DocumentTemplateSet = {
   name: 'Claw',
   description: 'Sharp, evolving agent with retractable claws that grip onto identity and purpose',
   tags: ['personality', 'evolving', 'autonomous'],
-  templates: [SOUL_DOCUMENT, IDENTITY_DOCUMENT, AGENT_DOCUMENT],
+  templates: [AGENT_DOCUMENT, BOOTSTRAP_DOCUMENT, IDENTITY_DOCUMENT, SOUL_DOCUMENT],
 };
 
 // Re-export individual templates for external use
-export { AGENT_DOCUMENT, IDENTITY_DOCUMENT, SOUL_DOCUMENT };
+export { AGENT_DOCUMENT, BOOTSTRAP_DOCUMENT, IDENTITY_DOCUMENT, SOUL_DOCUMENT };
