+++ packages/database/package.json
@@ -14,6 +14,7 @@
     "test:server-db": "vitest run --config vitest.config.server.mts"
   },
   "dependencies": {
+    "@lobechat/agent-templates": "workspace:*",
     "@lobechat/business-const": "workspace:*",
     "@lobechat/const": "workspace:*",
     "@lobechat/conversation-flow": "workspace:*",
