+++ src/server/modules/Mecha/ContextEngineering/index.ts
@@ -59,6 +59,7 @@ export const serverMessagesEngine = async ({
   historySummary,
   formatHistorySummary,
   knowledge,
+  agentDocuments,
   skillsConfig,
   toolsConfig,
   capabilities,
@@ -103,6 +104,7 @@ export const serverMessagesEngine = async ({
       fileContents: knowledge?.fileContents,
       knowledgeBases: knowledge?.knowledgeBases,
     },
+    agentDocuments,
 
     // Messages
     messages,
